- move collectRules to CountBasedOrderingStrategy
