+++ drools-core/src/main/java/org/drools/core/addon/NoopOrderingStrategy.java
@@ -15,17 +15,18 @@
 
 package org.drools.core.addon;
 
+import java.util.Collection;
 import java.util.List;
-import java.util.Set;
+import java.util.Map;
 
+import org.drools.core.definitions.InternalKnowledgePackage;
 import org.drools.core.spi.AlphaNodeFieldConstraint;
 import org.drools.core.spi.ObjectType;
-import org.kie.api.definition.rule.Rule;
 
 public class NoopOrderingStrategy implements AlphaNodeOrderingStrategy {
 
     @Override
-    public void analyzeAlphaConstraints(Set<Rule> ruleSet) {
+    public void analyzeAlphaConstraints(Map<String, InternalKnowledgePackage> pkgs, Collection<InternalKnowledgePackage> newPkgs) {
         // no op
     }
